fix(git): classify HTTP 5xx errors as auth errors with actionable guidance
HTTP 500/502/503 from GitHub during git clone often indicates stale
credentials from Git Credential Manager rather than genuine server errors.
Extract error classification into git-errors.ts and treat 5xx responses
as auth errors so users get guidance to run 'gh auth setup-git'.
